+++ src/batch.ts
@@ -1,10 +1,8 @@
 import type { DopriControlParam } from "dopri";
 
-import type {
-    InterpolatedSolution,
-    OdinModelConstructable,
-    SeriesSet,
-} from "./model";
+import type { OdinModelConstructable } from "./model";
+import type { InterpolatedSolution, SeriesSet } from "./solution";
+
 import { UserType } from "./user";
 import { grid, gridLog, loop, whichMax, whichMin } from "./util";
 import { wodinRun } from "./wodin";
